Create muxTB
Prior to running muxTB, a small change must be made to the mux file. The number of bits of the input on line 4 and the output reg on line 8 should be changed from 32 bits to 8 bits (from [31:0] to [7:0]). The test bench was made to run using 8 bits instead of 32 because the main file would not compile using 32 bits due to size.